+++ data/ss1.demo.sh
@@ -43,20 +43,20 @@ function add_str1() {
 	     name="Stream one" \
 	     description="Description of Stream one" \
 	     variables:='[ "baz", "bam" ]' \
-	     style:='{"color":"green", "dashArray": "5,5"}' > /dev/null
+	     mapStyle:='{"color":"green", "dashArray": "5,5"}' > /dev/null
 }
 
 function add_str2() {
     echo "add stream str2"
 	http post ${GPDVIZ}/api/ss/${SS} strid=str2 \
 	     variables:='[ "foo", "bar" ]' \
-	     style:='{"color":"red", "radius": 14}' zOrder:=10 > /dev/null
+	     mapStyle:='{"color":"red", "radius": 14}' zOrder:=10 > /dev/null
 }
 
 function add_str3() {
     echo "add stream str3"
 	http post ${GPDVIZ}/api/ss/${SS} strid=str3 \
-	     style:='{"color":"blue"}' > /dev/null
+	     mapStyle:='{"color":"blue"}' > /dev/null
 }
 
 function add_str1_values() {
@@ -149,7 +149,7 @@ function add_str4_and_point() {
     strid=str4
 	http post ${GPDVIZ}/api/ss/${SS} strid=${strid} \
 	    variables:='[ "temperature" ]' \
-	    style:='{"color":"yellow", "radius": 10}' zOrder:=10 > /dev/null
+	    mapStyle:='{"color":"yellow", "radius": 10}' zOrder:=10 > /dev/null
 	timestamp="`date +%s`000"
 	read -r -d '' values <<-EOF
 		[{